conclusão do exercicio

+++ exercicio/src/main/java/br/com/exercico/model/Rota.java
@@ -7,13 +7,13 @@ public class Rota extends BaseModel implements Serializable {
 	private static final long serialVersionUID = 1L;
 	private String origem;
 	private String destino;
-	private Integer distancia;
+	private double distancia;
 	
 	public Rota (){
 		super();
 	}
 	
-	public Rota (String origem, String destino, Integer distancia){
+	public Rota (String origem, String destino, double distancia){
 		this.destino = destino;
 		this.origem = origem;
 		this.distancia = distancia;
@@ -31,12 +31,15 @@ public class Rota extends BaseModel implements Serializable {
 	public void setDestino(String destino) {
 		this.destino = destino;
 	}
-	public Integer getDistancia() {
+
+	public double getDistancia() {
 		return distancia;
 	}
-	public void setDistancia(Integer distancia) {
+
+	public void setDistancia(double distancia) {
 		this.distancia = distancia;
 	}
+
